% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cinfeRence-package.R
\docType{package}
\name{cinfeRence-package}
\alias{cinfeRence}
\alias{cinfeRence-package}
\title{cinfeRence: A hands on package for deliberate practice in applied causal inference}
\description{
Tutorials for Applied Causal Inference
}
\author{
\strong{Maintainer}: Alex Stephenson \email{alexander_stephenson@berkeley.edu}

}
\keyword{internal}
