% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learncausalinfeRence-package.R
\docType{package}
\name{learncausalinfeRence-package}
\alias{learncausalinfeRence}
\alias{learncausalinfeRence-package}
\title{learncausalinfeRence: What the Package Does (One Line, Title Case)}
\description{
Tutorials for Applied Causal Inference
}
\author{
\strong{Maintainer}: Alex Stephenson \email{alexander_stephenson@berkeley.edu}

}
\keyword{internal}
